requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
asgiref==3.11.1
attrs==26.1.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==47.0.0
defusedxml==0.7.1
Django==6.0.4
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
frozenlist==1.8.0
idna==3.13
multidict==6.7.1
oauthlib==3.3.1
propcache==0.4.1
psycopg2-binary==2.9.12
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
python3-openid==3.2.0
requests==2.33.1
requests-oauthlib==2.0.0
social-auth-app-django==5.8.0
social-auth-core==4.8.7
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
pytest==8.3.5
pytest-django==4.10.0
